gjweb view 调试

+++ testCaseXcode7iOS9/testCaseXcode7iOS9/sylar/section/viewController/YYZBug/subViewIndex/GJPullingTableView/PullingTableViewVCTL.m
@@ -98,7 +98,7 @@
 
 - (BOOL)tableView:(UITableView *)tableView shouldHighlightRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    NSLog(@"sylar :  selected = %d,%d", indexPath.section, indexPath.row);
+    NSLog(@"sylar :  selected = %ld,%ld", (long)indexPath.section, (long)indexPath.row);
     UITableViewCell *cell = [tableView cellForRowAtIndexPath:indexPath];
     [tableView bringSubviewToFront:cell];
     return NO;
